+++ src/func/call.rs
@@ -276,6 +276,7 @@ impl Engine {
                                             func: CallableFunction::Method {
                                                 func: Shared::new(f),
                                                 has_context,
+                                                is_pure: false,
                                             },
                                             source: None,
                                         })
@@ -285,6 +286,7 @@ impl Engine {
                                         func: CallableFunction::Method {
                                             func: Shared::new(f),
                                             has_context,
+                                            is_pure: true,
                                         },
                                         source: None,
                                     }),
@@ -372,8 +374,8 @@ impl Engine {
 
             let backup = &mut ArgBackup::new();
 
-            // Calling pure function but the first argument is a reference?
-            let swap = is_ref_mut && func.is_pure() && !args.is_empty();
+            // Calling non-method function but the first argument is a reference?
+            let swap = is_ref_mut && !func.is_method() && !args.is_empty();
 
             if swap {
                 // Clone the first argument
@@ -400,12 +402,11 @@ impl Engine {
                 .has_context()
                 .then(|| (self, name, src, &*global, pos).into());
 
-            let mut _result = if let Some(f) = func.get_plugin_fn() {
-                if !f.is_pure() && !args.is_empty() && args[0].is_read_only() {
-                    Err(ERR::ErrorNonPureMethodCallOnConstant(name.to_string(), pos).into())
-                } else {
-                    f.call(context, args)
-                }
+            let mut _result = if !func.is_pure() && !args.is_empty() && args[0].is_read_only() {
+                // If function is not pure, there must be at least one argument
+                Err(ERR::ErrorNonPureMethodCallOnConstant(name.to_string(), pos).into())
+            } else if let Some(f) = func.get_plugin_fn() {
+                f.call(context, args)
             } else if let Some(f) = func.get_native_fn() {
                 f(context, args)
             } else {
@@ -603,8 +604,11 @@ impl Engine {
             };
 
             if error {
-                let sig = self.gen_fn_call_signature(fn_name, args);
-                return Err(ERR::ErrorFunctionNotFound(sig, pos).into());
+                return Err(ERR::ErrorFunctionNotFound(
+                    self.gen_fn_call_signature(fn_name, args),
+                    pos,
+                )
+                .into());
             }
         }
 
@@ -814,12 +818,16 @@ impl Engine {
             // Handle obj.call(fn_ptr, ...)
             KEYWORD_FN_PTR_CALL => {
                 if call_args.is_empty() {
-                    let typ = self.map_type_name(target.type_name());
-                    return Err(self.make_type_mismatch_err::<FnPtr>(typ, fn_call_pos));
+                    return Err(self.make_type_mismatch_err::<FnPtr>(
+                        self.map_type_name(target.type_name()),
+                        fn_call_pos,
+                    ));
                 }
                 if !call_args[0].is_fnptr() {
-                    let typ = self.map_type_name(call_args[0].type_name());
-                    return Err(self.make_type_mismatch_err::<FnPtr>(typ, first_arg_pos));
+                    return Err(self.make_type_mismatch_err::<FnPtr>(
+                        self.map_type_name(call_args[0].type_name()),
+                        first_arg_pos,
+                    ));
                 }
 
                 // FnPtr call on object
@@ -903,8 +911,10 @@ impl Engine {
             }
             KEYWORD_FN_PTR_CURRY => {
                 if !target.is_fnptr() {
-                    let typ = self.map_type_name(target.type_name());
-                    return Err(self.make_type_mismatch_err::<FnPtr>(typ, fn_call_pos));
+                    return Err(self.make_type_mismatch_err::<FnPtr>(
+                        self.map_type_name(target.type_name()),
+                        fn_call_pos,
+                    ));
                 }
 
                 let mut fn_ptr = target.read_lock::<FnPtr>().expect("`FnPtr`").clone();
@@ -1035,8 +1045,10 @@ impl Engine {
                     self.get_arg_value(global, caches, scope, this_ptr.as_deref_mut(), arg)?;
 
                 if !arg_value.is_fnptr() {
-                    let typ = self.map_type_name(arg_value.type_name());
-                    return Err(self.make_type_mismatch_err::<FnPtr>(typ, arg_pos));
+                    return Err(self.make_type_mismatch_err::<FnPtr>(
+                        self.map_type_name(arg_value.type_name()),
+                        arg_pos,
+                    ));
                 }
 
                 let fn_ptr = arg_value.cast::<FnPtr>();
@@ -1121,8 +1133,10 @@ impl Engine {
                     self.get_arg_value(global, caches, scope, this_ptr.as_deref_mut(), first)?;
 
                 if !arg_value.is_fnptr() {
-                    let typ = self.map_type_name(arg_value.type_name());
-                    return Err(self.make_type_mismatch_err::<FnPtr>(typ, arg_pos));
+                    return Err(self.make_type_mismatch_err::<FnPtr>(
+                        self.map_type_name(arg_value.type_name()),
+                        arg_pos,
+                    ));
                 }
 
                 let mut fn_ptr = arg_value.cast::<FnPtr>();
@@ -1493,17 +1507,18 @@ impl Engine {
                 self.call_script_fn(global, caches, scope, None, environ, f, args, true, pos)
             }
 
+            Some(f) if !f.is_pure() && args[0].is_read_only() => {
+                // If function is not pure, there must be at least one argument
+                Err(ERR::ErrorNonPureMethodCallOnConstant(fn_name.to_string(), pos).into())
+            }
+
             Some(f) if f.is_plugin_fn() => {
                 let f = f.get_plugin_fn().expect("plugin function");
                 let context = f
                     .has_context()
                     .then(|| (self, fn_name, module.id(), &*global, pos).into());
-                if !f.is_pure() && !args.is_empty() && args[0].is_read_only() {
-                    Err(ERR::ErrorNonPureMethodCallOnConstant(fn_name.to_string(), pos).into())
-                } else {
-                    f.call(context, args)
-                        .and_then(|r| self.check_data_size(r, pos))
-                }
+                f.call(context, args)
+                    .and_then(|r| self.check_data_size(r, pos))
             }
 
             Some(f) if f.is_native() => {