+++ CHANGELOG.md
@@ -11,12 +11,18 @@ Buf fixes
 
 * `is_shared` is a reserved keyword and is now handled properly (e.g. it cannot be the target of a function pointer).
 * Re-optimizing an AST via `optimize_ast` with constants now works correctly for closures. Previously the hidden `Share` nodes are not removed and causes variable-not-found errors during runtime if the constants are not available in the scope.
+* Expressions such as `(v[0].func()).prop` now parse correctly.
 
 New features
 ------------
 
 * It is now possible to require a specific _type_ to the `this` pointer for a particular script-defined function so that it is called only when the `this` pointer contains the specified type.
 * `is_def_fn` is extended to support checking for typed methods, with syntax `is_def_fn(this_type, fn_name, arity)`
+
+Enhancements
+------------
+
+* `Engine::is_symbol_disabled` is added to test whether a particular keyword/symbol is disabled.
 
 
 Version 1.13.0